Social Image Recognition/Code/requirements.txt
fastapi==0.111.0
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
requests==2.31.0
scikit_learn
tqdm==4.66.2
uvicorn==0.30.1
google-generativeai==0.5.2
regex==2024.4.28
pillow==10.3.0
